chore: proxy image preview in frontend (#11957)
