+++ blocksuite/affine/blocks/image/src/adapters/index.ts
@@ -1,4 +1,3 @@
 export * from './html.js';
 export * from './markdown.js';
-export * from './middleware.js';
 export * from './notion-html.js';
