+++ blocksuite/affine/blocks/image/src/index.ts
@@ -3,7 +3,6 @@ export * from './commands';
 export * from './edgeless-clipboard-config';
 export * from './image-block';
 export * from './image-edgeless-block';
-export { ImageProxyService } from './image-proxy-service';
 export * from './image-service';
 export * from './image-spec';
 export * from './styles';
