+++ blocksuite/affine/shared/src/adapters/middlewares/index.ts
@@ -3,6 +3,7 @@ export * from './copy';
 export * from './doc-link';
 export * from './file-name';
 export * from './paste';
+export * from './proxy';
 export * from './replace-id';
 export * from './surface-ref-to-embed';
 export * from './title';
